Checkpoint: moved shell_interactions.py under core/
