+++ manage_arkime/cdk_interactions/cdk_client.py
@@ -2,7 +2,7 @@ import logging
 import re
 from typing import Dict, List
 
-import shell_interactions as shell
+import core.shell_interactions as shell
 from aws_interactions.aws_environment import AwsEnvironment
 import cdk_interactions.cdk_exceptions as exceptions
 import core.constants as constants
