+++ manage_arkime/core/shell_interactions.py
